Cumulus acquirable grids in pqact
Python script for gridded files picked up on ldm and processed
with pqact.  Python 3.8 install and requirements added to the
LDM Docker image.  The script requres a file to transfer to s3,
an acquirable type (db acquirable) and a bucket (cwbi-data-develop
or cwbi-data-stable).  Multiple buckets can be used.  Environment
variables define the stable and develop buckets and URLs.